FIPS-compliant Dockerfile updates
This commit modifies the Dockerfile to have the rabbitmq-cluster-operator
to be built from ubi9 go-toolset and ubi-minimal images so to have
the necessary dependencies to be FIPS compliant.

It also squashes the build steps by removing the intermediate step
where we create the respective user/group.

Set FIPS flag in operator csv